+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 4.6.0
+
+#### Breaking
+- Updated to [Braze iOS SDK 4.5.1+](https://github.com/Appboy/appboy-ios-sdk/blob/master/CHANGELOG.md#451).
+
+#### Fixed
+- Fixed a bug where birthdays were not being logged properly from `identify` calls.
+
+#### Added
+- Adds `Other`, `Unknown`, `Not Applicable`, and `Prefer Not to Say` options for user gender.
+
 ## 4.5.0
 
 **Note:** This version does not include Carthage support. We are revisiting our Carthage approach and will reintroduce it in an upcoming version.